requirements.txt
Flask~=2.2.2
flask-restful~=0.3.9
boto3~=1.26.46
python-dotenv~=0.21.0
Werkzeug~=2.2.2